feat: change code to oop
