+++ navigation/lqr/nodes/lqr_node.py
@@ -5,19 +5,8 @@ lqr handler to integrate the following modules:
     - lqr_optimize_track
 """
 import rospy
-import numpy as np
-import matplotlib.pyplot as plt
 from nav_msgs.msg import Path
-from lqr import (
-    prepTrack,
-    Track,
-    SolverMatrices,
-    setupMatrices,
-    optimizeMinCurve,
-    createRaceLine,
-    pathToNumpy,
-    addWidth,
-)
+from lqr import Track
 
 TRACK_WIDTH = rospy.get_param("/navigation/lqr/handler/track_width")
 SAFETY_MARGIN = rospy.get_param("/navigation/lqr/handler/safety_margin")
@@ -27,52 +16,31 @@ def main() -> None:
     """
     Main Function of the lqrHandler
     """
-    referenceTrack = np.array([])
-    solverMatrices = SolverMatrices()
     rospy.init_node("lqr")
     rospy.loginfo("Waiting for Message")
     message = rospy.wait_for_message("/waypoints", Path, timeout=None)
     rospy.loginfo("Recieved Message")
 
-    trackBuffer = pathToNumpy(message)
-    referenceTrack = addWidth(trackBuffer)
-    trackClass = Track(referenceTrack)
-
-    # Setup the track
-    (
-        trackClass.smooth.track,
-        trackClass.smooth.normVectors,
-        trackClass.smooth.alpha,
-        trackClass.smooth.xCoeff,
-        trackClass.smooth.yCoeff,
-        trackClass.smooth.noPoints,
-        trackClass.smooth.noSplines,
-    ) = prepTrack(refTrack=trackClass.original)
-    setupMatrices(trackClass, solverMatrices)
-
-    # Calculate the optimal alpha values
-    trackClass.optimized.alpha = optimizeMinCurve(trackClass, solverMatrices)
-
-    # Calculate the track boundaries and raceline
-    raceLineMsg = Path()
-    raceLineMsg, raceLine, upperBound, lowerBound = createRaceLine(trackClass)
+    track = Track(message)
 
     # Publish the raceline
-    raceLinePub = rospy.Publisher("raceline", Path, queue_size=10)
-    raceLinePub.publish(raceLineMsg)
-    rospy.loginfo("Waypoints published")
+    raceLinePub = rospy.Publisher("/lqr/raceline", Path, queue_size=10)
+    upperBoundPub = rospy.Publisher("/lqr/upperBound", Path, queue_size=10)
+    lowerBoundPub = rospy.Publisher("/lqr/lowerBound", Path, queue_size=10)
 
-    # Plot
-    if rospy.get_param("/navigation/lqr/handler/plot") == 1:
-        plt.figure()
-        plt.plot(raceLine[:, 0], raceLine[:, 1])
-        plt.plot(trackClass.smooth.track[:, 0], trackClass.smooth.track[:, 1], "b--")
-        plt.plot(upperBound[:, 0], upperBound[:, 1], "k-")
-        plt.plot(lowerBound[:, 0], lowerBound[:, 1], "k-")
-        plt.grid()
-        plotArea = plt.gca()
-        plotArea.set_aspect("equal", "datalim")
-        plt.show()
+    while not rospy.is_shutdown():
+        try:
+            rospy.loginfo("Publishing waypoints")
+            raceLinePub.publish(track.optimized.racelineMsg)
+            upperBoundPub.publish(track.optimized.bounds.upperBoundMsg)
+            lowerBoundPub.publish(track.optimized.bounds.lowerBoundMsg)
+            rospy.sleep(5)
+        except rospy.ROSInterruptException:
+            rospy.loginfo("Shutting down")
+            break
+        except KeyboardInterrupt:
+            rospy.loginfo("Shutting down")
+            break
 
 
 if __name__ == "__main__":
