+++ navigation/lqr/src/lqr/__init__.py
@@ -1,6 +1,7 @@
 # pylint: disable=all
 # mypy: ignore-errors
-from .lqrOptimizeTrack import *
-from .lqrPrepareTrack import *
-from .track import *
-from .lqrCreateRaceline import *
+from .trackOriginal import *
+from .trackSmooth import *
+from .solverMatrices import *
+from .trackOptimized import *
+from .trackHandler import *
